+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+HEAD
+---
+* initial settings selector box is big enough to see some fields
+
 0.9.0.2
 ---
 * Add autocommit toggle